Enable 'mixed' operators with other strong_typedefs on underlying type
Mixed operators, such as mixed_addition, work fine when the underlying type
and the other type are compatible.
But when the other type is a strong_typedef, and they are not convertible
(which is the whole point of using a strong_typedef), mixed operators do
not work.
This commit checks whether the other type is a strong typedef, and if so,
applies the operator on the underlying type, instead of the strong typedef
itself.

Two use cases:
- Mixing a strong typedef for some value (such as 'year', like 2018) and some
offset related to such values (such as 'year_offset', like +3). These types
should not be implicitly convertible, but you do want 'year' to support
mixed addition with 'year_offset' on the underlying (int) type.
- Mixing an amount (euro's) strong typedef with a percentage strong typedef.
These types are obviously not convertible, but
mixed_multiplication<amount, percentage> on the underlying doubles is desirable