Workaround for compiling with MSVC /permissive-
There is a bug in MSVC ( https://godbolt.org/z/CZZvY3 ) due to which this code doesn't compile with the /permissive- option. Without /permissive- it compiles but then other things like ranges-v3 need /permissive- to compile. The auto return type deduction feature was added in Visual Studio 2015 (_MSC_VER 1900), so as a workaround we can just omit the trailing return type in newer versions and it works.